JSON String Interning
Problem:  JSON assets are one of the largest memory uses of
Starbound. A huge amount of this memory usage is strings representing
JsonObject keynames. There are only about 14,000 unique key strings in
Starbound, and common ones are stored in memory tens of thousands of
times.

Solution: String interning stores a database of key strings and
replaces them with a 16 bit integer.  This cuts down on memory usage
by about 150MB in-game and makes many JSON operations like merges much
faster.